2.20: added "Distance to next point", "Distance to destination", HR metrics in % of maxHR, last lap pace in seconds per 100 meter and kilcalories as metrics (not for FR645, FR735xt, FR935 and Fenix 5[s]). Also an option is added to remove the color of the labels for each position (not for FR645, FR735xt, FR935 and Fenix 5[s]).
Added power metrics in % of FTP, a Max-power metric and added a 5 and 10
sec averaged power metric (not for FR245m, FR645, FR645m, FR735xt, FR935
and Fenix 5[s]).

Added option to get power alerts for lap power instead of 3 sec averaged
power and added also an extra vibration when outside of the predefined
powerzone, which could make it discernible from being under the zone
only by the vibrations. (not for FR245m, FR645 and FR645m).